+++ ui-packages/packages/management-console-webapp/src/channel/ProcessList/ProcessListQueries.ts
@@ -15,11 +15,12 @@
  */
 
 import { ApolloClient } from 'apollo-client';
-import { SortBy, ProcessInstanceFilter } from '@kogito-apps/process-list';
 import {
   BulkProcessInstanceActionResponse,
   OperationType,
-  ProcessInstance
+  ProcessInstance,
+  SortBy,
+  ProcessInstanceFilter
 } from '@kogito-apps/management-console-shared';
 import {
   handleProcessAbort,
